Imporved extrinsics, still working on it for rosbags with rotated LiDAR

+++ src/main.cpp
@@ -1,3 +1,6 @@
+#include <tf2_ros/transform_broadcaster.h>
+#include <geometry_msgs/msg/transform_stamped.hpp>
+
 #include <as_lib/utils/Profiler.hpp>
 
 #include "Config.hpp"
@@ -9,7 +12,7 @@
 class Manager : public rclcpp::Node {
 
     Imu prevImu;
-    int firstStamp = -1;
+    double firstStamp = -1.;
     bool calibratedImu = false;
 
     Graph g;
@@ -55,7 +58,6 @@ class Manager : public rclcpp::Node {
         // conesPub = this->create_publisher<cat_msgs::msg::ConeArray>(cfg.topics.output.cones, 10);
         conesPub = this->create_publisher<visualization_msgs::msg::MarkerArray>(cfg.topics.output.cones, 10);
 
-        
     }
